Added network table and adjusted the alinemen
Move Network table to the vision class. Was originally in CrevoRobot.
However, the code was crashing every time we tried to use the
networktable. Discovered that its can't be called from a different
class.

Starting working on alinementToBoiler(void), Added so it doesn't run
through the process until it detected the boiler. Still needs more
calibration.